pyproject.toml
[project]
name = "tmp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "pycryptodome>=3.21.0",
    "requests>=2.32.3",
]
